refactor!: move GroupType to tile_group module
This makes `lib.rs` smaller and more organized.
