+++ src/main.rs
@@ -305,7 +305,8 @@ mod test {
     use mahc::fu::calculate_total_fu_value;
     use mahc::hand::error::HandErr;
     use mahc::hand::Hand;
-    use mahc::{GroupType, Suit};
+    use mahc::tile_group::GroupType;
+    use mahc::Suit;
 
     use super::*;
     #[test]
